2nd last update & password prompt deletion completed & search with name is done

+++ app/controllers/EmployeeController.php
@@ -210,14 +210,24 @@ class EmployeeController extends \BaseController {
 
 	public function destroy($id)
 	{
-		try{
+
+		$allInput = Input::all();
+		$data = Input::get('password');
+		$rules['password'] = 'required';
+		$validator = Validator::make($allInput,$rules);
+
+		if($validator->fails()){
+			return $this->errorResponse("Password Required", 400);
+		}
+
+		if(Hash::check($data,Auth::user()->password)) {
+			// return 'succes';
 			$user_id = Profile::find($id)->user_id;
 			User::destroy($user_id);
 
-			return Redirect::route('employee.index')->with('success','Employee Deleted Successfully.');
-
-		}catch(Exception $ex){
-			return Redirect::route('employee.index')->with('error','Something went wrong.Try Again.');
+			return $this->response('Employee Deleted Successfully.', 201);
+		} else {
+			return $this->errorResponse('Password did not match', 400);
 		}
 	}
 
@@ -243,6 +253,25 @@ class EmployeeController extends \BaseController {
 						->with('status', $status)
 						->with('title', 'Info of')
 						->with('salary', $salary);
+	}
+
+	
+
+
+	// code for json response to 
+	private function response($message, $status = 200){
+		return Response::json([
+						'data' => $message,
+						'status_code' => $status
+						
+					],$status);
+	}
+	private function errorResponse($message, $status = 400){
+		return Response::json([
+						'error' => $message,
+						'status_code' => $status
+						
+					], $status);
 	}
 
 }
